IMU fuctions added
This is the first compiling version for Arduino in which the functions
of the IMU are added. At the moment only the accelerometer and gyro
reaedings are used, so the other functionalities are still not
implemented, so this is still under construction. At this moment it
compiles but it hasn't been tested on the board yet.

Issues for Intel Galileo board:
- Variables called "time", such as the ones used in "Telemetry.cpp" had
to be changed to another name since it seems to be in conflict with
something else, still didn't figure out with what.
- Trying to write directly a "0" into the serial port is getting an
error message. Lines in "Telemetry.cpp" such as "Serial.write(0)" had to
be changed to "Serial.write( byte(0) )". The error we had was "call of
overloaded 'write(0)' is amiguous"