Monitoring processing time
- Modified the Move mode code in Arduino. The space bar now sends a
Short Command (instead of a Move mode byte) to stop all the motors.

- Added the calculation of the time (last, maximum and average) between
two calls to the function IMU.getMotion6 to know the average sampling
time for accelerometer and gyroscope data. Aimed to be less than 1ms,
which is the sampling time for the Accelerometer in the IMU. Also added
these times to the telemetry and are being displayed on the terminal
window in Eclipse on the computer side.

- Arranged the code in BXCommandMode.cpp